refactor: removing components from the app folder to the src folder and changing the page pat
