--- src/app/components/BackButton/BackButtonProps.d.ts
@@ -1,3 +1,0 @@
-export type BackButtonProps = {
-    href: string;
-};
