--- src/app/components/Button/ButtonProps.d.ts
@@ -1,4 +1,0 @@
-export type ButtonProps = {
-    title: string;
-    href: string;
-};
