--- src/app/components/CustomLink/CustomLinkProps.d.ts
@@ -1,3 +1,0 @@
-export type CustomLinkProps = {
-    text: string;
-};
