--- src/app/components/Form/FormProps.d.ts
@@ -1,4 +1,0 @@
-export type FormProps = {
-    label: string;
-    placeholder: string;
-};
