--- src/app/components/Submit/SubmitProps.d.ts
@@ -1,5 +1,0 @@
-export type SubmitProps = {
-    title: string;
-    type: "submit";
-    href: string;
-};
