--- src/app/components/Text/TextProps.d.ts
@@ -1,5 +1,0 @@
-export type TextProps = {
-    h1: string;
-    h2: string;
-    p: string;
-};
